requirements.txt
aiohttp==2.3.7
async-timeout==2.0.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
et-xmlfile==1.0.1
fake-useragent==0.1.8
idna==2.6
jdcal==1.3
multidict==3.3.2
openpyxl==2.4.9
python-dotenv==0.9.1
requests==2.18.4
telepot==12.5
urllib3==1.22
yarl==0.18.0
